+++ src/itemRadio/ItemRadio.tsx
@@ -105,7 +105,10 @@ class ItemRadio extends Component<ItemRadioProps> {
           rightTitle={data}
           rightTitleDisplay={TextDisplayType.SUBHEADERSTRONG}
           rightBody={dataInfo}
-          /* TODO: BBC-7415 fix a11y issue */
+          /* No a11y issue here
+            - The input is well wrapped with the label
+            - The linter can't access the complex components implementation
+          */
           // eslint-disable-next-line jsx-a11y/label-has-associated-control
           tag={<label />}
           rightAddon={radio}