+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Unreleased
 
+- **[FIX]** Remove DOM Warning on `MainContent` component
+- **[FIX]** Fix markup `Item` tag label by replacing `div` children with `span`
 - **[BREAKING CHANGE]** Import `icons` as default instead of named export for a better code splitting
+  [...]
 
 # v25.0.0 (02/04/2020)